# perf: move screen focus to LHNOptionsList